renaming modules to support py3.7
